+++ src/eamonnsullivan/github_pr_lib.clj
@@ -1,15 +1,12 @@
 (ns eamonnsullivan.github-pr-lib
   (:require [clj-http.client :as client]
-            [clojure.string :as string]
             [clojure.data.json :as json]
             [clojure.java.io :as io]))
 
 (def github-url "https://api.github.com/graphql")
-(def ^:dynamic *search-page-size* 50)
 
 (def get-repo-id-query (slurp (io/resource "graphql/get-repo-id-query.graphql")))
 (def create-pull-request-mutation (slurp (io/resource "graphql/create-pull-request-mutation.graphql")))
-(def search-for-pr-id-query (slurp (io/resource "graphql/search-for-pr-id-query.graphql")))
 (def update-pull-request-mutation (slurp (io/resource "graphql/update-pull-request-mutation.graphql")))
 (def mark-ready-for-review-mutation (slurp (io/resource "graphql/mark-ready-for-review-mutation.graphql")))
 (def add-comment-mutation (slurp (io/resource "graphql/add-comment-mutation.graphql")))
@@ -18,7 +15,6 @@
 (def reopen-pull-request-mutation (slurp (io/resource "graphql/reopen-pull-request-mutation.graphql")))
 (def merge-pull-request-mutation (slurp (io/resource "graphql/merge-pull-request-mutation.graphql")))
 (def pull-request-query (slurp (io/resource "graphql/pull-request-query.graphql")))
-(def search-for-issue-comment-id (slurp (io/resource "graphql/search-for-issue-comment-id-query.graphql")))
 
 (defn request-opts
   "Add the authorization header to the http request options."
@@ -26,9 +22,40 @@
   {:ssl? true :headers {"Authorization" (str "bearer " access-token)}})
 
 (defn http-post
-  "Make a POST request to a url with body payload and request options."
+  "Make a POST request to a url with payload and request options."
   [url payload opts]
   (client/post url (merge {:content-type :json :body payload} opts)))
+
+(defn http-get
+  "Make a GET request to a url, with options"
+  [access-token url options]
+  (client/get url (merge {:username access-token} options)))
+
+(defn get-pull-request-node-id
+  "Get the node id of a pull request using the v3 REST api, optionally
+  filtered by state (\"open\", \"closed\", or the default of \"all\")"
+  ([access-token owner repo-name pullnum]
+   (get-pull-request-node-id access-token owner repo-name pullnum "all"))
+  ([access-token owner repo-name pull-number state]
+   (let [url (str "https://api.github.com/repos/" owner
+                  "/" repo-name
+                  "/pulls/" pull-number)
+         response (http-get access-token url {:throw-exceptions false
+                                              :accept "application/vnd.github.v3+json"
+                                              :query-params {"state" state}})
+         body (json/read-str (:body response) :key-fn keyword)]
+     (:node_id body))))
+
+(defn get-comment-node-id
+  "Get the node id of a pull request comment using the v3 REST API."
+  [access-token owner repo comment-number]
+  (let [url (str "https://api.github.com/repos/" owner
+                 "/" repo
+                 "/issues/comments/" comment-number)
+        response (http-get access-token url {:throw-exceptions false
+                                             :accept "application/vnd.github.v3+json"})
+        body (json/read-str (:body response) :key-fn keyword)]
+    (:node_id body)))
 
 (defn make-graphql-post
   "Make a GraphQL request to Github using the provided query/mutation
@@ -64,16 +91,16 @@
       nil)))
 
 (defn parse-comment-url
-  "Get the full comment url and pull request url from an issue comment URL."
+  "Get the comment number and pull request url from an issue comment URL."
   [comment-url]
-  (let [matches (re-matches #"(https://github.com/)?([^/]*)/([^/]*)/pull/([0-9]*)#(issuecomment-[0-9]*)" comment-url)
+  (let [matches (re-matches #"(https://github.com/)?([^/]*)/([^/]*)/pull/([0-9]*)#issuecomment-([0-9]*)" comment-url)
         [_ _ owner name number comment] matches]
     (if (and (not-empty owner)
              (not-empty name)
              (not-empty number)
              (not-empty comment))
       {:pullRequestUrl (format "https://github.com/%s/%s/pull/%s" owner name number)
-       :issueComment (format "#%s" comment)}
+       :issueComment comment}
       nil)))
 
 (defn get-repo-id
@@ -92,76 +119,38 @@
          :repository
          :id))))
 
-(defn get-page-of-pull-requests
-  "Get a page of pull requests, optionally (and by default) filtered by
-  those with a status of open."
-  ([access-token owner name page-size cursor]
-   (get-page-of-pull-requests access-token owner name page-size cursor true))
-  ([access-token owner name page-size cursor open?]
-   (let [variables {:owner owner :name name :first page-size :after cursor}]
-     (if open?
-       (make-graphql-post access-token search-for-pr-id-query (merge {:states ["OPEN"]} variables))
-       (make-graphql-post access-token search-for-pr-id-query (merge {:states ["OPEN" "CLOSED" "MERGED"]} variables))))))
-
 (defn get-pull-request-id
   "Find the unique ID of a pull request on the repository at the
   provided url. Set must-be-open? to true to filter the pull requests
   to those with a status of open. Returns nil if not found."
-  [access-token url must-be-open?]
-  (let [repo (parse-repo url)
-        prnum (pull-request-number url)
-        owner (:owner repo)
-        name (:name repo)
-        pull-request-url (string/lower-case (format "https://github.com/%s/%s/pull/%s" owner name prnum))
-        page (get-page-of-pull-requests access-token owner name *search-page-size* nil must-be-open?)]
-    (loop [page page
-           prs []]
-      (let [pageInfo (-> page :data :repository :pullRequests :pageInfo)
-            has-next (:hasNextPage pageInfo)
-            cursor (:endCursor pageInfo)
-            pull-requests (-> page :data :repository :pullRequests :nodes)
-            prs (concat prs pull-requests)]
-        (if-not has-next
-          (:id (first (filter #(= (:url %) pull-request-url) prs)))
-          (recur (get-page-of-pull-requests access-token owner name *search-page-size* cursor must-be-open?)
-                 prs))))))
-
-(defn get-page-of-issue-comments
-  "Get a page of issues comments on a particular pull request"
-  [access-token pull-request-id page-size cursor]
-  (let [variables {:pullRequestId pull-request-id :first page-size :after cursor}]
-    (make-graphql-post access-token search-for-issue-comment-id variables)))
+  ([access-token url]
+   (get-pull-request-id access-token url false))
+  ([access-token url must-be-open?]
+   (let [repo (parse-repo url)
+         prnum (pull-request-number url)
+         owner (:owner repo)
+         name (:name repo)]
+     (get-pull-request-node-id access-token owner name prnum (if must-be-open? "open" "all")))))
 
 (defn get-issue-comment-id
   "Find the unique ID of an issue comment on a pull request. Returns nil if not found."
   [access-token comment-url]
-  (let [prurl (:pullRequestUrl (parse-comment-url comment-url))]
-    (if prurl
-      (let [prnum (get-pull-request-id access-token prurl false)
-            page (get-page-of-issue-comments access-token prnum *search-page-size* nil)]
-        (loop [page page
-               comments []]
-          (let [pageInfo (-> page :data :node :comments :pageInfo)
-                has-next (:hasNextPage pageInfo)
-                cursor (:endCursor pageInfo)
-                nodes (-> page :data :node :comments :nodes)
-                comments (concat comments nodes)]
-            (if-not has-next
-              (:id (first (filter #(= (:url %) comment-url) comments)))
-              (recur (get-page-of-issue-comments access-token prnum *search-page-size* cursor)
-                     comments)))))
-      nil)))
+  (let [repo (parse-repo comment-url)
+        owner (:owner repo)
+        name (:name repo)
+        comment (:issueComment (parse-comment-url comment-url))]
+    (get-comment-node-id access-token owner name comment)))
 
-(defn get-open-pr-id
-  "Find the unique ID of an open pull request. Returns nil of none are found."
-  ([access-token pull-request-url]
-   (get-pull-request-id access-token pull-request-url true)))
 
 (defn get-pull-request-info
-  "Find some info about a pull request."
+  "Find some info about a pull request.
+
+  Available properties: :id, :title, :body, :baseRefOid,
+  :headRefOid, :permalink, :author (:login, :url), :closed,
+  :isDraft, :merged, :mergeable (MERGEABLE, CONFLICTING or UNKNOWN),
+  :number, :repository (:id, :url), :state (CLOSED, MERGED or OPEN)."
   [access-token pull-request-url]
-  (let [pr-id (or (get-open-pr-id access-token pull-request-url)
-                  (get-pull-request-id access-token pull-request-url false))]
+  (let [pr-id (get-pull-request-id access-token pull-request-url)]
     (when pr-id
       (-> (make-graphql-post access-token pull-request-query {:pullRequestId pr-id})
           :data
@@ -172,7 +161,7 @@
   ([access-token url mutation]
    (modify-pull-request access-token url mutation nil))
   ([access-token url mutation variables]
-   (let [pr-id (or (get-open-pr-id access-token url) (get-pull-request-id access-token url false))]
+   (let [pr-id (get-pull-request-id access-token url)]
      (when pr-id
        (let [merged-variables (merge variables {:pullRequestId pr-id})]
          (make-graphql-post access-token mutation merged-variables))))))