+++ test/eamonnsullivan/github_pr_lib_test.clj
@@ -4,28 +4,25 @@
             [clojure.string :as string]
             [clojure.data.json :as json]))
 
-(def repo-id-response-success (slurp "./resources/test-fixtures/repo-response-success.json"))
-(def repo-id-response-failure (slurp "./resources/test-fixtures/repo-response-failure.json"))
-(def create-pr-response-success (slurp "./resources/test-fixtures/create-pr-response-success.json"))
-(def create-pr-response-failure (slurp "./resources/test-fixtures/create-pr-response-failure.json"))
-(def first-pr-search-page (slurp "./resources/test-fixtures/first-page-pr.json"))
-(def second-pr-search-page (slurp "./resources/test-fixtures/second-page-pr.json"))
-(def first-comment-search-page (slurp "./resources/test-fixtures/first-page-comments.json"))
-(def second-comment-search-page (slurp "./resources/test-fixtures/second-page-comments.json"))
-(def update-pr-success (slurp "./resources/test-fixtures/update-pr-success.json"))
-(def update-pr-failure (slurp "./resources/test-fixtures/update-pr-failure.json"))
-(def mark-ready-success (slurp "./resources/test-fixtures/mark-ready-success.json"))
-(def mark-ready-failure (slurp "./resources/test-fixtures/mark-ready-failure.json"))
-(def add-comment-success (slurp "./resources/test-fixtures/add-comment-success.json"))
-(def add-comment-failure (slurp "./resources/test-fixtures/add-comment-failure.json"))
-(def edit-comment-success (slurp "./resources/test-fixtures/edit-comment-success.json"))
-(def edit-comment-failure (slurp "./resources/test-fixtures/edit-comment-failure.json"))
-(def close-pull-request-success (slurp "./resources/test-fixtures/close-pull-request-success.json"))
-(def close-pull-request-failure (slurp "./resources/test-fixtures/close-pull-request-failure.json"))
-(def reopen-pull-request-success (slurp "./resources/test-fixtures/reopen-pull-request-success.json"))
-(def reopen-pull-request-failure (slurp "./resources/test-fixtures/reopen-pull-request-failure.json"))
-(def merge-pull-request-success (slurp "./resources/test-fixtures/merge-pull-request-success.json"))
-(def merge-pull-request-failure (slurp "./resources/test-fixtures/merge-pull-request-failure.json"))
+(def repo-id-response-success (slurp "./test/eamonnsullivan/fixtures/repo-response-success.json"))
+(def repo-id-response-failure (slurp "./test/eamonnsullivan/fixtures/repo-response-failure.json"))
+(def create-pr-response-success (slurp "./test/eamonnsullivan/fixtures/create-pr-response-success.json"))
+(def create-pr-response-failure (slurp "./test/eamonnsullivan/fixtures/create-pr-response-failure.json"))
+(def update-pr-success (slurp "./test/eamonnsullivan/fixtures/update-pr-success.json"))
+(def update-pr-failure (slurp "./test/eamonnsullivan/fixtures/update-pr-failure.json"))
+(def mark-ready-success (slurp "./test/eamonnsullivan/fixtures/mark-ready-success.json"))
+(def mark-ready-failure (slurp "./test/eamonnsullivan/fixtures/mark-ready-failure.json"))
+(def add-comment-success (slurp "./test/eamonnsullivan/fixtures/add-comment-success.json"))
+(def add-comment-failure (slurp "./test/eamonnsullivan/fixtures/add-comment-failure.json"))
+(def edit-comment-success (slurp "./test/eamonnsullivan/fixtures/edit-comment-success.json"))
+(def edit-comment-failure (slurp "./test/eamonnsullivan/fixtures/edit-comment-failure.json"))
+(def close-pull-request-success (slurp "./test/eamonnsullivan/fixtures/close-pull-request-success.json"))
+(def close-pull-request-failure (slurp "./test/eamonnsullivan/fixtures/close-pull-request-failure.json"))
+(def reopen-pull-request-success (slurp "./test/eamonnsullivan/fixtures/reopen-pull-request-success.json"))
+(def reopen-pull-request-failure (slurp "./test/eamonnsullivan/fixtures/reopen-pull-request-failure.json"))
+(def merge-pull-request-success (slurp "./test/eamonnsullivan/fixtures/merge-pull-request-success.json"))
+(def merge-pull-request-failure (slurp "./test/eamonnsullivan/fixtures/merge-pull-request-failure.json"))
+(def pull-request-properties (slurp "./test/eamonnsullivan/fixtures/pull-request-properties.json"))
 
 (deftest test-get-repo-id
   (with-redefs [sut/http-post (fn [_ _ _] {:body repo-id-response-success})]
@@ -145,6 +142,15 @@
     (is (= nil (sut/pull-request-number "something else")))
     (is (= nil (sut/pull-request-number "https://github/owner/name/pull")))))
 
+
+(deftest test-get-pull-request-id
+  (with-redefs [sut/http-get (fn [_ _ _] {:body "{\"node_id\": \"a-node-id\"}"})]
+    (testing "finds the node id in the body"
+      (is (= "a-node-id" (sut/get-pull-request-id "secret" "https://github.com/owner/name/pull/1")))))
+  (with-redefs [sut/http-get (fn [_ _ _] {:body "{}"})]
+    (testing "returns nil on failure"
+      (is (= nil (sut/get-pull-request-id "secret" "https://github.com/owner/name/pull/2"))))))
+
 (defn fake-paging-post
   [page1 page2]
   (fn [_ payload _]
@@ -154,38 +160,22 @@
         {:body page2}
         {:body page1}))))
 
-(deftest test-get-pr-id
-  (with-redefs [sut/http-post (fake-paging-post first-pr-search-page second-pr-search-page)]
-    (testing "finds pull request id"
-      (is (= "MDExOlB1bGxSZXF1ZXN0MTU5NjI3ODc2" (sut/get-open-pr-id "secret" "https://github.com/eamonnsullivan/something/pull/2")))
-      (is (= "MDExOlB1bGxSZXF1ZXN0MTYwOTE1ODA0" (sut/get-open-pr-id "secret" "https://github.com/eamonnsullivan/something/pull/5"))))
-    (testing "finds pull request id on subsequent pages"
-      (is (= "MDExOlB1bGxSZXF1ZXN0MTU5NjI3ODc2" (sut/get-open-pr-id "secret" "https://github.com/eamonnsullivan/something/pull/7"))))))
-
 (deftest test-get-issue-comment-id
-  (with-redefs [sut/http-post (fake-paging-post first-comment-search-page second-comment-search-page)]
-    (testing "finds comment id"
-      (is (= "MDEyOklzc3VlQ29tbWVudDcwMjA3NjE0Ng=="
-             (sut/get-issue-comment-id
-              "secret"
-              "https://github.com/eamonnsullivan/github-pr-lib/pull/4#issuecomment-702076146")))
-      (is (= "MDEyOklzc3VlQ29tbWVudDcwMjA5MjY4Mg=2"
-             (sut/get-issue-comment-id
-              "secret"
-              "https://github.com/eamonnsullivan/github-pr-lib/pull/4#issuecomment-702092683")))
-      (is (= nil
-             (sut/get-issue-comment-id
-              "secret"
-              "https://github.com/eamonnsullivan/github-pr-lib/pull/4#issuecomment-70207614999"))))))
+  (with-redefs [sut/http-get (fn [_ _ _] {:body "{\"node_id\": \"a-node-id\"}"})]
+    (testing "finds the node id in the body"
+      (is (= "a-node-id" (sut/get-issue-comment-id "secret" "https://github.com/owner/name/pull/1#issuecomment-1")))))
+  (with-redefs [sut/http-get (fn [_ _ _] {:body "{}"})]
+    (testing "returns nil on failure"
+      (is (= nil (sut/get-issue-comment-id "secret" "https://github.com/owner/name/pull/1#issuecomment-1"))))))
 
 (deftest test-update-pull-request
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body update-pr-success})]
     (testing "updates a pull request"
       (is (= "https://github.com/eamonnsullivan/github-pr-lib/pull/3" (sut/update-pull-request "secret"
                                                                                                "https://github.com/eamonnsullivan/github-pr-lib/pull/3"
                                                                                                {:title "A new title" :body "A new body"})))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body update-pr-failure})]
     (testing "Throws exception on update error"
       (is (thrown-with-msg? RuntimeException #"Could not resolve to a node with the global id of 'invalid-id'"
@@ -193,12 +183,12 @@
                                                      "https://github.com/eamonnsullivan/github-pr-lib/pull/3"
                                                      {:title "A new title" :body "A new body"}))))))
 (deftest test-mark-ready-for-review
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body mark-ready-success})]
     (testing "marks pull request as ready for review"
       (is (= "https://github.com/eamonnsullivan/github-pr-lib/pull/3" (sut/mark-ready-for-review "secret"
                                                                                                  "https://github.com/eamonnsullivan/github-pr-lib/pull/3")))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body mark-ready-failure})]
     (testing "Throws exception on error"
       (is (thrown-with-msg? RuntimeException #"Could not resolve to a node with the global id of 'invalid-id'"
@@ -206,13 +196,13 @@
                                                        "https://github.com/eamonnsullivan/github-pr-lib/pull/3"))))))
 
 (deftest test-add-pull-request-comment
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body add-comment-success})]
     (testing "adds comment to pull request"
       (is (= "https://github.com/eamonnsullivan/github-pr-lib/pull/4#issuecomment-702076146" (sut/add-pull-request-comment "secret"
                                                                                                                            "https://github.com/eamonnsullivan/github-pr-lib/pull/4"
                                                                                                                            "This is a comment.")))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body add-comment-failure})]
     (testing "Throws exception on error"
       (is (thrown-with-msg? RuntimeException #"Could not resolve to a node with the global id of 'invalid'"
@@ -241,12 +231,12 @@
                              "This is an updated comment."))))))
 
 (deftest test-close-pull-request
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body close-pull-request-success})]
     (testing "closes a pull request"
       (is (= "https://github.com/eamonnsullivan/github-pr-lib/pull/4" (sut/close-pull-request "secret"
                                                                                               "https://github.com/eamonnsullivan/github-pr-lib/pull/4")))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body close-pull-request-failure})]
     (testing "Throws exception on error"
       (is (thrown-with-msg? RuntimeException #"Could not resolve to a node with the global id of 'invalid'"
@@ -254,14 +244,12 @@
                                                     "https://github.com/eamonnsullivan/github-pr-lib/pull/4"))))))
 
 (deftest test-reopen-pull-request
-  (with-redefs [sut/get-open-pr-id (fn [_ _] nil)
-                sut/get-pull-request-id (fn [_ _ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body reopen-pull-request-success})]
     (testing "reopens a pull request"
       (is (= "https://github.com/eamonnsullivan/github-pr-lib/pull/4" (sut/reopen-pull-request "secret"
                                                                                                "https://github.com/eamonnsullivan/github-pr-lib/pull/4")))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] nil)
-                sut/get-pull-request-id (fn [_ _ _] "some-id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some-id")
                 sut/http-post (fn [_ _ _] {:body reopen-pull-request-failure})]
     (testing "Throws exception on error"
       (is (thrown-with-msg? RuntimeException #"Could not resolve to a node with the global id of 'invalid'"
@@ -276,7 +264,7 @@
       (is (not= nil (:expectedHeadRef variables))))))
 
 (deftest test-merge-pull-request
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some id")
                 sut/get-pull-request-info (fn [_ _] {:headRefOid "commit-id"})
                 sut/http-post (fn [_ _ _] {:body merge-pull-request-success})]
     (testing "merges a pull request"
@@ -284,7 +272,7 @@
                                                                                               "https://github.com/eamonnsullivan/github-pr-lib/pull/4"
                                                                                               {:title "a commit" :body "some description"
                                                                                                :author-email "someone@somewhere.com"})))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some id")
                 sut/get-pull-request-info (fn [_ _] {:headRefOid "commit-id"})
                 sut/http-post (fn [_ _ _] {:body merge-pull-request-failure})]
     (testing "Throws exception on error"
@@ -293,7 +281,7 @@
                                                     "https://github.com/eamonnsullivan/github-pr-lib/pull/4"
                                                     {:title "a commit" :body "some description"
                                                      :author-email "someone@somewhere.com"})))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some id")
                 sut/get-pull-request-info (fn [_ _] nil)
                 sut/http-post (fn [_ _ _] {:body merge-pull-request-failure})]
     (testing "Throws exception if the pull request can't be found"
@@ -302,7 +290,7 @@
                                                     "https://github.com/eamonnsullivan/github-pr-lib/pull/4"
                                                     {:title "a commit" :body "some description"
                                                      :author-email "someone@somewhere.com"})))))
-  (with-redefs [sut/get-open-pr-id (fn [_ _] "some id")
+  (with-redefs [sut/get-pull-request-id (fn [_ _] "some id")
                 sut/get-pull-request-info (fn [_ _] {:headRefOid "commit-id"})
                 sut/http-post (make-fake-post nil merge-pull-request-success nil assert-merge-payload-defaults)]
     (testing "merges a pull request"
@@ -316,8 +304,22 @@
            (:pullRequestUrl (sut/parse-comment-url "https://github.com/eamonnsullivan/github-pr-lib/pull/4#issuecomment-702092682"))))
     (is (= "https://github.com/eamonnsullivan/github-pr-lib/pull/4"
            (:pullRequestUrl (sut/parse-comment-url "eamonnsullivan/github-pr-lib/pull/4#issuecomment-702092682"))))
-    (is (= "#issuecomment-702092682"
+    (is (= "702092682"
            (:issueComment (sut/parse-comment-url "eamonnsullivan/github-pr-lib/pull/4#issuecomment-702092682")))))
   (testing "return nil when a pull request url can't be found"
     (is (= nil
            (:pullRequestUrl (sut/parse-comment-url "https://news.bbc.co.uk"))))))
+
+(deftest test-get-pull-request-info
+  (with-redefs [sut/http-post (fn [_ _ _] {:body pull-request-properties})
+                sut/get-pull-request-id (fn [_ _] "some-id")]
+    (let [response (sut/get-pull-request-info "secret" "owner/name/pull/1")]
+      (is (= "A test title" (:title response)))
+      (is (= "With a body" (:body response)))
+      (is (= "johnsmith" (-> response :author :login)))
+      (is (= "https://github.com/owner/name" (-> response :repository :url)))
+      (is (= false (:isDraft response)))
+      (is (= "MERGED" (:state response)))
+      (is (= true (:merged response)))
+      (is (= "UNKNOWN" (:mergeable response)))
+      (is (= 9 (:number response))))))